add title desluggify feature
A common use case of routes is to provide a slug of a title or name of document. This leads to breadcrumb in a form:

    level 1 > My-awesome-title > level 3

What we usually want is for that to look like:

    level 1 > My Awesome Title > level 3

With this patch, user can specify a slug character that was used in the slug. So if we have:

    title: ':param',
    slug: '-'

All the '-' characters in the title will be changed into ' ' and potentially capitalized.